+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Remove account from the system - all BBs [#335](https://github.com/rokwire/core-building-block/issues/335)
 
+## [1.40.0] - 2024-05-09
+### Changed
+- Add "admin_app_access_roles" to the system POST and PUT app_orgs [#712](https://github.com/rokwire/core-building-block/issues/712) 
+
+## [1.39.0] - 2024-04-30
+### Changed
+- Admin app account creation update [#710](https://github.com/rokwire/core-building-block/issues/710)
+
 ## [1.38.1] - 2024-03-05
 ### Fixed
 - Fix v2 create accounts [#704](https://github.com/rokwire/core-building-block/issues/704)
@@ -445,7 +453,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Users authentication polish [#155](https://github.com/rokwire/core-building-block/issues/155)
 - Optimise the Mongo DB collections indexes usage [#146](https://github.com/rokwire/core-building-block/issues/146)
 
-[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.38.1...HEAD
+[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.39.0...HEAD
+[1.39.0]: https://github.com/rokwire/core-building-block/compare/v1.38.1...v1.39.0
 [1.38.1]: https://github.com/rokwire/core-building-block/compare/v1.38.0...v1.38.1
 [1.38.0]: https://github.com/rokwire/core-building-block/compare/v1.37.2...v1.38.0
 [1.37.2]: https://github.com/rokwire/core-building-block/compare/v1.37.1...v1.37.2